backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
anthropic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
openai